requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
pipenv==2021.5.29
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
virtualenv-clone==0.5.4
whitenoise==5.2.0
